requirements.txt
anthropic
pydantic
python-dotenv
rich
loguru 